feat(onboarding): guided post-verification setup flow
Route a just-verified user through a two-step onboarding sequence
(appearance preferences, then reference photos) before the coach chat,
instead of dropping them straight on /chat. Both steps are skippable;
nothing here gates the coach.

Refactor the two reused forms by extraction so the account page is
unchanged:
- Split AppearanceSelector into AppearanceFields (bare selectors + shared
  helpers) and a thin account-page wrapper.
- Split ReferenceImageManager into ReferenceImageUploader (upload core,
  with optional heading/intro) and a thin card wrapper.

Add the /onboarding route and page, and divert the verify-email redirect
to /onboarding. No backend changes.
